+++ internal/server/server.go
@@ -2,8 +2,12 @@ package server
 
 import (
 	"context"
-	"fmt"
+	"net"
+	"net/http"
+	"strconv"
+	"time"
 
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/rs/zerolog"
 
 	"github.com/authorizerdev/authorizer/internal/graphql"
@@ -18,6 +22,8 @@ type Config struct {
 	HTTPPort int
 	// Port number to serve Metrics requests on
 	MetricsPort int
+	// MetricsHost is the bind address for the dedicated /metrics listener.
+	MetricsHost string
 }
 
 // Dependencies for a server
@@ -42,19 +48,38 @@ type server struct {
 	Dependencies *Dependencies
 }
 
-// Run the server until the given context is canceled
+// Run the server until the given context is canceled.
+// The main HTTP server (Gin) and the Prometheus /metrics server always run as separate listeners.
 func (s *server) Run(ctx context.Context) error {
-	// Create new router
 	ginRouter := s.NewRouter()
-	// Start the server
+	httpAddr := net.JoinHostPort(s.Config.Host, strconv.Itoa(s.Config.HTTPPort))
 	go func() {
-		s.Dependencies.Log.Info().Str("host", s.Config.Host).Int("port", s.Config.HTTPPort).Msg("Starting HTTP server")
-		err := ginRouter.Run(s.Config.Host + ":" + fmt.Sprintf("%d", s.Config.HTTPPort))
-		if err != nil {
+		s.Dependencies.Log.Info().Str("addr", httpAddr).Msg("Starting HTTP server")
+		if err := ginRouter.Run(httpAddr); err != nil {
 			s.Dependencies.Log.Error().Err(err).Msg("HTTP server failed")
 		}
 	}()
-	// Wait until context closed
+
+	mux := http.NewServeMux()
+	mux.Handle("/metrics", promhttp.Handler())
+	metricsAddr := net.JoinHostPort(s.Config.MetricsHost, strconv.Itoa(s.Config.MetricsPort))
+	metricsSrv := &http.Server{
+		Addr:    metricsAddr,
+		Handler: mux,
+	}
+	go func() {
+		s.Dependencies.Log.Info().Str("addr", metricsAddr).Msg("Starting metrics server")
+		if err := metricsSrv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
+			s.Dependencies.Log.Error().Err(err).Msg("Metrics server failed")
+		}
+	}()
+	go func() {
+		<-ctx.Done()
+		shCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+		_ = metricsSrv.Shutdown(shCtx)
+	}()
+
 	<-ctx.Done()
 	return nil
 }